django_project/telegram_bot/requirements.txt
-i https://pypi.org/simple
django==2.2.5
pytz==2019.1
sqlparse==0.3.0
uwsgi==2.0.18
dnspython==1.16.0
requests==2.22.0

